+++ crates/board/src/platform/protocol.rs
@@ -41,4 +41,7 @@ pub trait Api {
 
     /// Disables events from being triggered.
     fn disable() -> Result<(), Error>;
+
+    /// Handles vendor-specific requests.
+    fn vendor(request: &[u8]) -> Result<Box<[u8]>, Error>;
 }