+++ crates/cli/CHANGELOG.md
@@ -4,15 +4,22 @@
 
 ### Minor
 
+- Add `platform-rpc` for vendor-specific platform RPCs
+- Implement `platform-reboot` command
+- Add `--timeout` for platform protocol timeout
+- Implement `applet-rpc` command
+- Add `--serial` and `WASEFIRE_SERIAL` to select the platform
+- Implement `platform-list` command
 - Support creating, building, and testing a Rust applet
 - Support generating a shell completion file
 
 ### Patch
 
+- Publish LICENSE file
 - Update dependencies
 - Depend on features of dependencies explicitly
 - Use common Wasefire lints
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 0 -->
+<!-- Increment to skip CHANGELOG.md test: 1 -->