+++ crates/prelude/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Minor
 
+- Add `platform::serial()` to get the platform serial
+- Migrate `platform::version()` to new applet API
 - Add `rpc::Rpc` trait for `platform::protocol::RpcProtocol`
 - Add the `store::{keys,clear}()` functions
 - Add the `platform::protocol` module and the `api-platform-protocol` feature for applet RPC
@@ -21,6 +23,7 @@
 
 ### Patch
 
+- Publish LICENSE file
 - Move lints to `Cargo.toml`, use common Wasefire lints, and make most dependencies optional
 - Update dependencies